Simplify ROM chip logic.
